pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "runner"
version = "0.1.0"
[tool.poetry.dependencies]
dblpy = "^0.3.4"
"discord.py" = "^1.2"
flask = "^1.1"
jishaku = "^1.17"
psutil = "^5.6"
python = "^3.8"
python-dotenv = "^0.13.0"
[tool.poetry.dev-dependencies]
